refactor(#78): 로비 시작 가능 여부 계산 정책 분리
